third_party/blink/web_tests/jquery/dimensions-expected.txt

ALERT: 58 tests of 58 passed, 0 failed.
